Rename IOI contest format to ICS3U contest format
