+++ judge/contest_format/__init__.py
@@ -2,5 +2,6 @@ from judge.contest_format.atcoder import AtCoderContestFormat
 from judge.contest_format.bonuses import BonusesContestFormat
 from judge.contest_format.default import DefaultContestFormat
 from judge.contest_format.ecoo import ECOOContestFormat
+from judge.contest_format.ics3u import ICS3UContestFormat
 from judge.contest_format.ioi import IOIContestFormat
 from judge.contest_format.registry import choices, formats
